refactor(deck): rename deck components for clarity
- components/deck.vue → components/deck/deck-thumbnail.vue
- components/deck/deck-preview.vue → components/deck/deck-design-preview.vue
- delete unused components/popovers/deck-preview.vue + stale locale keys
- update imports, template tags, data-testids, and test stubs
