+++ src/locales/en-us.json
@@ -325,7 +325,6 @@
   "deck-view.mode-view.new-card": "New Card",
   "deck-view.card-editor.list-item.front-placeholder": "Front",
   "deck-view.card-editor.list-item.back-placeholder": "Back",
-  "deck-preview-popover.study": "Study",
   "deck-view.actions.study-all": "Study <span>all</span> cards",
   "deck-view.actions.edit-cards": "Edit Cards",
   "deck-view.actions.stop-editing": "Stop Editing",
@@ -339,7 +338,6 @@
   "deck-view.card-importer.import-button": "Import",
   "deck-view.card-importer.save-button.idle": "Save",
   "deck-view.card-importer.save-button.saving": "Saving…",
-  "deck-preview-popover.image-alt": "Deck Image preview",
   "deck-view.tabs.card-view": "View",
   "deck-view.tabs.edit-cards": "Edit",
   "deck-view.tabs.import-cards": "Import",
